Reduce dependencies
Installation should be as simple as possible. As this is one of the
first programs I install on a new system, the friction of having to
install cmake/pkg-config/openssl just to get my promptline compiled is a
pain. Remove expensive dependency (libgit2) and replace with manual
parsing of git info. While we're in the mindset of reducing
dependencies, remove dependency on anyhow and fall back to more bare
bones style rust error handling